Awesome timeline.
Moved some properties of nav in .main-nav, so that other nav
elements can be created reusing some properties.
A lot of css and js to make the timeline work.
New color variable: @cloud (for the kind-of-white).
New .scroll-to class on .btn button-links (automatic smooth
scrolling).
Import a new FlatUI js file for the radio button.
Refactored the coffee file: better organization, a lot of
refactoring.